Undo `TestEnv` move.
